Remove OculusHandController in favor of agnostics MRHandController
