+++ XRTK.Oculus/Packages/com.xrtk.oculus/Controllers/Hands.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 099f9f26548043741903fd3506b6862e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
