+++ XRTK.Oculus/Packages/com.xrtk.oculus/Controllers/Hands/OculusHandDataConverter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c93983b816bea5a4c986651c57040666
+guid: 15984a1d9794fdc4fa8fe2748bc18839
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
